1. ByteArrayCheckContains追加 2. コマンド引数でDecimal使用箇所をdoubleに変更

+++ EmVerif/Core/Script/Command/DegreeCheckCommand.cs
@@ -11,16 +11,16 @@ namespace EmVerif.Core.Script.Command
     class DegreeCheckCommand : IEmVerifCommand
     {
         private string _FormulaStr;
-        private Decimal _ExpValueMax;
-        private Decimal _ExpValueMin;
+        private double _ExpValueMax;
+        private double _ExpValueMin;
         private string _Message;
         private Regex _VarNameRegex = new Regex(@"(?<VarName>[a-zA-Z_][\w\[\]]*)");
         private DataTable _Dt = new DataTable();
 
         public DegreeCheckCommand(
             string inFormula,
-            Decimal inExpValueMax,
-            Decimal inExpValueMin,
+            double inExpValueMax,
+            double inExpValueMin,
             string inMessage
         )
         {
@@ -55,7 +55,7 @@ namespace EmVerif.Core.Script.Command
 
         public void Finally(ControllerState inState)
         {
-            Decimal checkValue = ConvertFormula(inState, _FormulaStr) % 360;
+            double checkValue = ConvertFormula(inState, _FormulaStr) % 360;
             string result;
 
             if (
@@ -90,7 +90,7 @@ namespace EmVerif.Core.Script.Command
             LogManager.Instance.Set("\t" + _ExpValueMin.ToString() + @" <= " + _FormulaStr + @" <= " + _ExpValueMax.ToString() + @"⇒" + result);
         }
 
-        private Decimal ConvertFormula(ControllerState inState, string inOrgFormula)
+        private double ConvertFormula(ControllerState inState, string inOrgFormula)
         {
             var varNameMatches = _VarNameRegex.Matches(inOrgFormula);
             string resultStr = inOrgFormula;
@@ -112,7 +112,7 @@ namespace EmVerif.Core.Script.Command
                 }
             }
 
-            return Convert.ToDecimal(_Dt.Compute(resultStr, ""));
+            return Convert.ToDouble(_Dt.Compute(resultStr, ""));
         }
     }
 }
